Move watch history logging from the system service
Now the system service is only responsible for inserting raw data when the TV is
tuned to a new channel (watch start time) or the current input session is
released (watch end time), which minimizes database operations in the system
service. On the other hand, it is now responsibility of TvProvider to fill up
missing program information and consolidate the log entries.

Bug: 14791151, Bug: 15921205, Bug: 15988715, Bug: 16831555
Change-Id: I468c35f2e120448f184380be4c8899a8e5e55d36